Create README.md
This is a project i did last week in order to showcase what i'd learn in StemLabNg. While the place i'm doing my internship(stemlab) provides me and my fellow interns a
comprehensive tech knowledge that help us build, lunch and scale, i figure out that there is still some certain challenges we face, so i come up with a solution to that which we may encounter. The problems are like; competion, interns adoption, content quality, interns feedback, funding/recruiting and so on. Amazingly, i come up with a very good and reliable solutions in order for us the interns to be successful not only in the stemlab but also in the tech industry.